[make_tree.c] after split into 2 files (make_tree_redir_node)
